requirements.txt
Babel==2.5.3
Django==2.0.2
django-phonenumber-field==2.0.0
django-tel==0.3
mod-wsgi==4.5.24
phonenumberslite==8.9.0
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
